# server/swagger.yaml
openapi: 3.0.3
info:
  title: EventKnit API
  description: |
    EventKnit is a comprehensive ticketing platform API that enables event management, ticket sales, payments, and real-time communications.

    ## Authentication
    Most endpoints require JWT authentication. Include the access token in the Authorization header:
    ```
    Authorization: Bearer <your_access_token>
    ```

    ## Base URL
    All API routes are prefixed with `/api/v1`

    ## Rate Limiting
    API requests are rate-limited to prevent abuse. Rate limits vary by endpoint type.
  version: 1.0.0
  contact:
    name: EventKnit Support
    email: support@eventknit.com
  license:
    name: ISC
servers:
  - url: http://localhost:3001/api/v1
    description: Development server
  - url: https://api.eventknit.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from login/registration
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
        data:
          type: object
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5

    # Career Inquiry Schemas
...
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        category:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        location:
          type: object
          properties:
            address:
              type: string
            city:
              type: string
            state:
              type: string
            country:
              type: string
            latitude:
              type: number
            longitude:
              type: number
        eventType:
          type: string
          enum: [PHYSICAL, VIRTUAL, HYBRID]
        status:
          type: string
          enum: [DRAFT, PENDING_APPROVAL, APPROVED, REJECTED, CANCELLED, COMPLETED]
        isPublished:
          type: boolean
        capacity:
          type: integer
        coverImage:
          type: string
        organizerId:
          type: string
          format: uuid
        organizer:
          $ref: '#/components/schemas/User'
        tickets:
          type: array
          items:
            $ref: '#/components/schemas/TicketType'
        registrationCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    BadgeElement:
      type: object
      required:
        - id
        - type
        - content
        - x
        - y
        - width
        - height
        - zIndex
      properties:
        id:
          type: string
          description: Unique element identifier
        type:
          type: string
          enum: [text, image, qr, shape, logo, barcode, ribbon]
          description: Element type
        content:
          type: string
          description: Element content (can include template variables like {{fullName}})
        x:
          type: number
          description: X position in millimeters
        y:
          type: number
          description: Y position in millimeters
        width:
          type: number
          description: Width in millimeters
        height:
          type: number
          description: Height in millimeters
        fontSize:
          type: number
          nullable: true
          description: Font size for text elements
        fontFamily:
          type: string
          nullable: true
          description: Font family for text elements
        fontWeight:
          type: string
          enum: [normal, bold, light]
          nullable: true
        textAlign:
          type: string
          enum: [left, center, right]
          nullable: true
        color:
          type: string
          nullable: true
          description: Text or foreground color (hex format)
        backgroundColor:
          type: string
          nullable: true
          description: Background color (hex format)
        borderRadius:
          type: number
          nullable: true
          description: Border radius in pixels
        opacity:
          type: number
          nullable: true
          minimum: 0
          maximum: 1
        rotation:
          type: number
          nullable: true
          description: Rotation angle in degrees
        zIndex:
          type: integer
          description: Layer order (higher = on top)
        isVisible:
          type: boolean
          nullable: true
        isLocked:
          type: boolean
          nullable: true
        ribbonIcon:
          type: string
          enum: [crown, star, check-circle, award, badge, none]
          nullable: true
          description: Icon for ribbon elements
        ribbonGradient:
          type: string
          nullable: true
          description: CSS gradient string for ribbon background
    BadgeTemplate:
      type: object
      required:
        - id
        - name
        - width
        - height
        - sizePreset
        - orientation
        - backgroundColor
        - elements
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: Template name
        description:
          type: string
          nullable: true
          description: Template description
        width:
          type: number
          description: Badge width in millimeters
          example: 101.6
        height:
          type: number
          description: Badge height in millimeters
          example: 76.2
        sizePreset:
          type: string
          enum: [4x3, 3.5x2.25, 4x6, A6, custom]
          description: Preset badge size
        orientation:
          type: string
          enum: [portrait, landscape]
        backgroundColor:
          type: string
          description: Background color (hex format)
          example: "#ffffff"
        backgroundImage:
          type: string
          nullable: true
          description: URL to background image (Cloudinary)
        elements:
          type: array
          items:
            $ref: '#/components/schemas/BadgeElement'
        isDefault:
          type: boolean
          description: Whether this is the default template
        isActive:
          type: boolean
          description: Whether the template is active
        organizerId:
          type: string
          format: uuid
          nullable: true
        eventId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Unauthorized access
    ForbiddenError:
      description: User does not have permission to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: You do not have permission to perform this action
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Validation failed
            errors:
              - field: email
                message: Valid email is required
paths:
...
  /events:
    get:
      tags:
        - Events
      summary: Get all events
      description: Retrieve a list of all public events with optional filters
      operationId: getEvents
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, PENDING_APPROVAL, APPROVED, REJECTED, CANCELLED, COMPLETED]
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: eventType
          in: query
          schema:
            type: string
            enum: [PHYSICAL, VIRTUAL, HYBRID]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Events retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          events:
                            type: array
                            items:
                              $ref: '#/components/schemas/Event'
                          pagination:
                            $ref: '#/components/schemas/PaginationMeta'

    post:
      tags:
        - Events
      summary: Create a new event
      description: Create a new event (requires ORGANIZER role or higher)
      operationId: createEvent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - description
                - startDate
                - endDate
                - category
                - eventType
              properties:
                title:
                  type: string
                description:
                  type: string
                shortDescription:
                  type: string
                category:
                  type: string
                startDate:
                  type: string
                  format: date-time
                endDate:
                  type: string
                  format: date-time
                location:
                  type: object
                  properties:
                    address:
                      type: string
                    city:
                      type: string
                    state:
                      type: string
                    country:
                      type: string
                eventType:
                  type: string
                  enum: [PHYSICAL, VIRTUAL, HYBRID]
                capacity:
                  type: integer
                coverImage:
                  type: string
                tickets:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      description:
                        type: string
                      price:
                        type: number
                      quantity:
                        type: integer
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Event'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /organizer/events:
    get:
      tags:
        - Organizers
      summary: Get organizer events
      description: Get all events created by the organizer
      operationId: getOrganizerEvents
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Events retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          events:
                            type: array
                            items:
                              $ref: '#/components/schemas/Event'
                          pagination:
                            $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /badge-templates:
    get:
      tags:
        - Service Point
      summary: Get badge templates
      description: Retrieve badge templates with optional filters
      operationId: getBadgeTemplates
      security:
        - BearerAuth: []
      parameters:
        - name: organizerId
          in: query
          schema:
            type: string
          description: Filter by organizer ID
        - name: eventId
          in: query
          schema:
            type: string
          description: Filter by event ID
        - name: isDefault
          in: query
          schema:
            type: boolean
          description: Filter by default status
        - name: isActive
          in: query
          schema:
            type: boolean
          description: Filter by active status
        - name: search
          in: query
          schema:
            type: string
          description: Search by name or description
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Templates retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          templates:
                            type: array
                            items:
                              $ref: '#/components/schemas/BadgeTemplate'
                          total:
                            type: integer
                          page:
                            type: integer
                          limit:
                            type: integer
                          totalPages:
                            type: integer
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Service Point
      summary: Create badge template
      description: Create a new badge template
      operationId: createBadgeTemplate
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                description:
                  type: string
                width:
                  type: number
                  default: 101.6
                height:
                  type: number
                  default: 76.2
                sizePreset:
                  type: string
                  enum: [4x3, 3.5x2.25, 4x6, A6, custom]
                  default: 4x3
                orientation:
                  type: string
                  enum: [portrait, landscape]
                  default: landscape
                backgroundColor:
                  type: string
                  default: "#ffffff"
                elements:
                  type: array
                  items:
                    $ref: '#/components/schemas/BadgeElement'
                  default: []
                isDefault:
                  type: boolean
                  default: false
                organizerId:
                  type: string
                eventId:
                  type: string
      responses:
        '201':
          description: Template created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          template:
                            $ref: '#/components/schemas/BadgeTemplate'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /careers:
    post:
      tags:
        - Careers
      summary: Submit career inquiry
      description: |
        Submit a career inquiry by providing an email address.
        An automated email will be sent with instructions for the candidate to share their story.
        Rate limited to one submission per email per 24 hours.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: candidate@example.com
                  description: Candidate's email address
                source:
                  type: string
                  maxLength: 100
                  example: linkedin
                  description: Optional tracking source
      responses:
        '201':
          description: Career inquiry submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Thank you for your interest! Check your email for next steps.
                  data:
                    type: object
                    properties:
                      email:
                        type: string
                        example: candidate@example.com
        '400':
          description: Invalid email format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Careers
      summary: List career inquiries (Admin)
      description: Get all career inquiries with pagination and filtering. Requires admin privileges.
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, CONTACTED, RESPONDED, REVIEWING, REJECTED, HIRED]
          description: Filter by inquiry status
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: Items per page
      responses:
        '200':
          description: List of career inquiries
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      inquiries:
                        type: array
                        items:
                          $ref: '#/components/schemas/CareerInquiry'
                      pagination:
                        $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /cart:
    get:
      tags:
        - Cart
      summary: Get current cart
      description: |
        Retrieves the current user's active shopping cart.
        Supports both authenticated users (via JWT) and guest users (via session cookie or X-Cart-Session header).
        Returns null if no active cart exists.
      operationId: getCart
      security: []
      parameters:
        - name: X-Cart-Session
          in: header
          description: Session ID for guest users (alternative to cookie)
          schema:
            type: string
      responses:
        '200':
          description: Cart retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      cart:
                        $ref: '#/components/schemas/CartReservation'
                      message:
                        type: string
                        example: No active cart
              examples:
                activeCart:
                  summary: Active cart with items
                  value:
                    success: true
                    data:
                      cart:
                        id: "550e8400-e29b-41d4-a716-446655440000"
                        sessionId: "session-123"
                        status: "ACTIVE"
                        expiresAt: "2024-12-01T10:30:00Z"
                        items:
                          - id: "item-1"
                            eventId: "evt-1"
                            ticketType: "General Admission"
                            quantity: 2
                            unitPrice: 50.00
                            inventoryLocked: false
                noCart:
                  summary: No active cart
                  value:
                    success: true
                    data:
                      cart: null
                      message: "No active cart"

    post:
      tags:
        - Cart
      summary: Create a new cart
      description: |
        Creates a new shopping cart. If an active cart already exists for the session/user,
        returns the existing cart instead of creating a new one.

        Cart expires after 8 minutes of inactivity.
      operationId: createCart
      security: []
      parameters:
        - name: X-Cart-Session
          in: header
          description: Session ID for guest users (alternative to cookie)
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                fingerprint:
                  type: string
                  description: Browser fingerprint for fraud detection
      responses:
        '201':
          description: Cart created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      cart:
                        $ref: '#/components/schemas/CartReservation'

    delete:
      tags:
        - Cart
      summary: Abandon cart
      description: |
        Abandons the current cart and releases any locked inventory.
        Clears the cart session cookie.
      operationId: abandonCart
      security: []
      parameters:
        - name: X-Cart-Session
          in: header
          description: Session ID for guest users
          schema:
            type: string
      responses:
        '200':
          description: Cart abandoned successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Cart abandoned successfully
        '404':
          $ref: '#/components/responses/NotFoundError'
...
security:
  - BearerAuth: []